requirements.txt
exa-py 
openai 
faiss-cpu 
python-dotenv 
numpy
langsmith 
langchain
python-dotenv
langchain-openai